build: abstract cookie manager
Ref: metriport/metriport-internal#1195
